# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight REST relay for managing TR-069 devices via the GenieACS NBI. v2.2.0 adds 25 new operational endpoints across 4 phases (CPE lifecycle, NOC support, customer self-service, GenieACS metadata) on top of the v2.1.0 baseline. 100% main-package coverage, 40/40 endpoints end-to-end verified on real ZTE F670L V9.0.10P1N12A via VPN lab (sessions 5i + 5j 2026-04-15).\nSee CHANGELOG.md [2.2.0] for full release notes and upstream genieacs-stack v1.3.1 blocker for customer factory-reset workflows.",
        "title": "GenieACS Relay API",
        "contact": {
            "name": "Cepat Kilat Teknologi",
            "url": "https://github.com/Cepat-Kilat-Teknologi/genieacs-relay"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/genieacs",
    "paths": {
...
        "/wlan/available/{ip}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns available WLAN slots for creating new WiFi networks based on device capability",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WLAN"
                ],
                "summary": "Get available WLAN slots",
                "parameters": [
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.AvailableWLANResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/wlan/create/{wlan}/{ip}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new WLAN on an available slot with advanced configuration options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WLAN"
                ],
                "summary": "Create new WLAN",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2",
                        "description": "WLAN ID (1-4 for 2.4GHz, 5-8 for 5GHz)",
                        "name": "wlan",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "WLAN creation parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateWLANRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.WLANCreateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/wlan/delete/{wlan}/{ip}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Disables an existing WLAN configuration. This effectively \"deletes\" the WLAN by disabling it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WLAN"
                ],
                "summary": "Delete/Disable WLAN",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2",
                        "description": "WLAN ID (1-8)",
                        "name": "wlan",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.WLANDeleteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/wlan/optimize/{wlan}/{ip}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Optimizes WLAN radio settings including channel, mode, bandwidth, and transmit power. Supports partial updates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WLAN"
                ],
                "summary": "Optimize WLAN radio settings",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "WLAN ID (1-8)",
                        "name": "wlan",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "WLAN optimization parameters (all fields optional)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.OptimizeWLANRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.WLANOptimizeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.CreateWLANRequest": {
            "type": "object",
            "properties": {
                "auth_mode": {
                    "description": "Authentication mode: Open, WPA, WPA2, WPA/WPA2",
                    "type": "string"
                },
                "encryption": {
                    "description": "Encryption mode: AES, TKIP, TKIP+AES",
                    "type": "string"
                },
                "hidden": {
                    "description": "Hide SSID (SSIDAdvertisementEnabled = false)",
                    "type": "boolean"
                },
                "max_clients": {
                    "description": "Maximum number of associated devices (1-64)",
                    "type": "integer"
                },
                "password": {
                    "description": "Password for the new WLAN (required for WPA/WPA2)",
                    "type": "string"
                },
                "ssid": {
                    "description": "SSID for the new WLAN",
                    "type": "string"
                }
            }
        },
...
        "main.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "data": {
                    "description": "Response payload (success) OR error detail (string/field list)"
                },
                "error_code": {
                    "description": "Machine-readable error code (VALIDATION_ERROR, NOT_FOUND, ...)",
                    "type": "string"
                },
                "request_id": {
                    "description": "Correlation ID from X-Request-ID header",
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" for 2xx, HTTP reason phrase for errors",
                    "type": "string"
                }
            }
        },
...
        "main.WLANCreateResponse": {
            "description": "WLAN creation response",
            "type": "object",
            "properties": {
                "auth_mode": {
                    "type": "string",
                    "example": "WPA2"
                },
                "band": {
                    "type": "string",
                    "example": "2.4GHz"
                },
                "device_id": {
                    "type": "string",
                    "example": "001141-F670L-ZTEGCFLN794B3A1"
                },
                "encryption": {
                    "type": "string",
                    "example": "AES"
                },
                "hidden": {
                    "type": "boolean",
                    "example": false
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "max_clients": {
                    "type": "integer",
                    "example": 10
                },
                "message": {
                    "type": "string",
                    "example": "WLAN creation submitted successfully"
                },
                "ssid": {
                    "type": "string",
                    "example": "GuestNetwork"
                },
                "wlan": {
                    "type": "string",
                    "example": "2"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication (required when MIDDLEWARE_AUTH=true)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}